requirements.txt
python-dotenv==0.19.2

pymongo~=3.12.3